Saving and Loading works now, atleast with the little manual tests I've run in REPL.
Had to use some workarounds to include game in COM, so now every Player has var game but it is now used in Human and it only has an invalid Game as a place holder.

also added a bunch of methods to player and game. Player: chageGame and returnGame. Changed .toString a little, now tells whether player is Human or COM.
Game: addPlayer atleast.